requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
PyJWT==2.1.0
drf-spectacular==0.22.1
Pillow==9.0.0
